Add support for local builds
This commit adds supports for local builds. In particular, a `getBuildPackageStats`  and `getBuildExportsSizes` now also accepts a path to a local build directory. This path is valid if the directory contains a package.json file.  `npm` supports local folder installations, so the main changes are related to how the `packageString` parameter is parsed into a name and version.  The temporary installPath directory now uses the `packageName` as opposed to the `packageString`.

When using `getBuildExportSizes`, `getExportAll` expects the context to be a substring of the install path, however, with local builds this is not the case because they are symlinked to the installation directory. The implication is that the resolver which is used to resolve the export path in  context, follows the link and returns the original local build path. To get around this, we pass the local build path (the packageString) as the context when we know it's a local package.