fix: local builds resolve dependencySizes properly
When `getPackageStats` is run, the expected output includes composite information for npm dependency. This information is outputted in the `dependencySizes` array.

Previous to this commit, this information was not getting calculated correctly when `getPackageStats` was given a local path to install. The issue was that npm installs local package directory structure than with a remote package.

Local:
`<installpath>/node_modules/package.json`
`<installpath>/node_modules/<packageName>` // this a symlink to the local package
`<installpath>/node_modules/<packageName>/package.json`
`<installpath>/node_modules/<packageName>/node_modules`

Remote:
`<installpath>/node_modules/package.json`
`<installpath>/node_modules/*` // All the packages (deps and the package itself) are install here.

The parsing logic in `getDirectorySizeTree` did not account for the extra `node_modules` in the module path and due to this was not calculating the dep tree correctly.