Use npm pack when installing local directories.
The behaviour of `npm install <packageName>` and `npm install <localPath>` is inconsistent. More specially, `npm` when operating on a local directory, `npm` just links the folder in `node_modules` as opposed to installing it and all of its dependencies in a flatten structure. This is important because `package-build-stats` makes this assumption when navigating the dependency tree.

Rather than trying to determine the nature of the node_modules ad hoc, it would be better to operate on a consistent structure. We can achieve this with in`npm` by first packing the local package folder, and running `install` against the packed gzipped file. It's also worth noting this is how `yarn` installs local package folders (they don't do the linking).

With this commit, `dependencySizes` will be correctly be calculated.