Add --ignore-scripts to npm pack method
A local package with a malicious script would be executed when packing.

This can be recreated by creating a new package with a `prepack` script and running `getPackageStats` with the local path.

```jsonc
// package.json
{
  "name": "maliciousPackage",
  "scripts": {
    "prepack": "send your env variables to me"
  }
}
```

An issue with no scripts is that any benevolent scripts wouldn't occur, such as compilation. But this would already not occur with non-local files.